### 2019-06-24 #1523 @whedon
[PRE REVIEW]: pyMARS: automatically reducing chemical kinetic models in Python
**Submitting author:** @kyleniemeyer (<a href="http://orcid.org/0000-0003-4425-7097">Kyle E Niemeyer</a>)
**Repository:** <a href="https://github.com/Niemeyer-Research-Group/pyMARS" target ="_blank">https://github.com/Niemeyer-Research-Group/pyMARS</a>
**Version:** v1.0.0
