### 2019-06-24 #1523 @katyhuff
Oh darn. I was going to suggest @kyleniemeyer as the editor... But I see he's the author.

Sure, I'll do it. @kyleniemeyer , do you have reviewer suggestions (obviously, folks who have no COI reviewing your paper).
