### 2019-06-24 #1523 @kyleniemeyer
Thanks @katyhuff! Here are some suggestions.

No conflicts from my domain:
- Prof. James Sutherland (@jcsutherland) or team members at Univ. Utah, like @michael-a-hansen
- Prof. Perrine Pepiot at Cornell (perrine.pepiot@cornell.edu)
- Prof. Francesco Contino at Vrije Universiteit Brussel (@fcontino)

From the JOSS pool:
- @ctdegroot
- @zbeekman
- @adam-m-jcbs
