### 2019-06-24 #1523 @katyhuff
@zbeekman : Zaak Beekman are you able to and interested in taking on this review?

* A JOSS review involves checking software submissions against a checklist of essential software features and details in the submitted paper. Additional information about reviewing for JOSS can be found [here.](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)
* This submission is in need of expertise in chemical kinetics as well as familiarity with the systems and languages involved (primarily python)
* Title: pyMARS: automatically reducing chemical kinetic models in Python
* Summary: Python-based (chemical kinetic) Model Automatic Reduction Software (pyMARS) implements multiple techniques for reducing the size and complexity of detailed chemical kinetic models.
* Article Proof: [pdf](https://github.com/openjournals/joss-papers/blob/joss.01523/joss.01523/10.21105.joss.01523.pdf)
* Submitting author: @kyleniemeyer (Kyle Niemeyer)
* Repository: https://github.com/Niemeyer-Research-Group/pyMARS
* Version: 1.0.0
