### 2019-06-24 #1523 @zbeekman
I'm interested but a bit short on time these days. I just finished on JOSS review (f90nml), and my portion of another is still in progress but nearly done.

Also, I don't know if this would be a COI, but my employer, [ParaTools Inc.](www.paratools.com), has a commercial product, [KPP Accelerated (KPPA)](http://www.paratools.com/kppa/), that may have some overlap of pyMARS, and, *might* be a competitive alternative. (I say might, because I have not yet dug into the specifics of pyMARS' capabilities, and I'm only familiar with some of KPPA's.) It is also a python based tool, although its main purpose is to generate fast C code that can interface with Fortran, C, etc. and run on accelerators like GPUs. I'm not sure if KPPA does any model reduction, or just graph-partitioning/matrix-ordering etc. optimizations.

All things being equal, this may not be considered a COI on JOSS' end, but it's possible that it would be my employer's end. If you cannot come up with two other reviewers, then please let me know and I can ask if reviewing this would be OK.
