### 2019-06-25 #1523 @fcontino
Sure and with great pleasure. I'll need some time thought because I have
many other review tasks on my plate right now. What would be an acceptable
deadline?

Thank you for involving me in this.

Best regards,

Francesco
