### 2019-06-29 #1523 @jcsutherland
I'm happy to provide a review, but it will take me at least a week.
--
James C. Sutherland
Professor of Chemical Engineering
The University of Utah
50 S. Central Campus Dr, 3290 MEB
Salt Lake City, UT 84112-9203
+1-801-585-1246
sutherland <https://sutherland.che.utah.edu/>.che.utah.edu
<https://sutherland.che.utah.edu>
<https://doodle.com/jamessutherland>
