### 2019-07-02 #1523 @zbeekman
@jcsutherland once the editor (@katyhuff) selects the reviewers via whedon commands, it will open a new review issue, with more instructions. So sit tight until that happens.

@katyhuff it looks like you have two reviewers now. Please LMK if you need me for anything, but I'm unsubscribing for now :wave:. @ mention me if you want to get my attention.
