### 2019-07-02 #1523 @katyhuff
@fcontino : We would like reviews to be completed in two weeks if possible. I hope that will be an acceptable deadline for you. Notably, the paper is quite short and since @kyleniemeyer knows the requirements for a JOSS submission quite well, we hope it won't take more than a few hours once you begin.
