### 2019-07-02 #1543 @whedon
[REVIEW]: pyMARS: automatically reducing chemical kinetic models in Python
**Submitting author:** @kyleniemeyer (<a href="http://orcid.org/0000-0003-4425-7097">Kyle E Niemeyer</a>)
**Repository:** <a href="https://github.com/Niemeyer-Research-Group/pyMARS" target ="_blank">https://github.com/Niemeyer-Research-Group/pyMARS</a>
**Version:** v1.1.0

@fcontino, @jcsutherland please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
