### 2019-07-02 #1523 @katyhuff
Ok folks, let's move over to #1543 where you will find guidance and a checklist regarding the review process.
