### 2019-07-05 9ecc85c
added check for conditions file when parsing command-line args
