### 2019-07-05 d4ba039
Fixed checking of saved ignition file contents
