### 2019-07-05 07735a6
Added Anaconda package badge to README
