### 2019-07-05 95f44c3
Update docs after building Travis build 89 of Niemeyer-Research-Group/pyMARS
The docs were built from the tag 'v1.0.0a9' against the commit
c6aa27b9e1512741717ecaa2659a51cd3a5a9782.

The Travis build that generated this commit is at
https://travis-ci.org/Niemeyer-Research-Group/pyMARS/jobs/554468859.

The doctr command that was run is

    /home/travis/miniconda/bin/doctr deploy --build-tags tags/v1.0.0a9
