+++ tags/v1.0.0a9/_sources/usage.rst.txt
@@ -1,0 +1,184 @@
+.. _sec_usage:
+
+=====
+Usage
+=====
+
+.. toctree::
+    :maxdepth: 1
+
+Once installed, pyMARS can be run as an executable. Change to the directory
+where your input files are located, then run::
+
+    pymars [options]
+
+The following options are available, and can also be seen by using the
+``-h`` or ``--help`` option:
+
+.. code-block:: bash
+
+     -h, --help:
+        show this help message and exit
+     -m, --model:
+        input model filename (e.g., mech.cti)
+     -e, --error:
+        Maximum error percentage for the reduced model
+     --method {DRG,DRGEP,PFA}
+        skeletal reduction method to use
+     --conditions:
+        File with list of autoignition initial conditions
+     --targets:
+        List of target species (e.g., "CH4 O2")
+     --retained_species:
+        List of non-target species to always retain (e.g., "N2 Ar")
+     --sensitivity_analysis:
+        Run sensitivity analysis after completing another method
+     --sensitivity_type {initial, greedy}
+        Sensitivity analysis method to use
+     --upper_threshold:
+        Upper threshold value used to determine species for sensitivity analysis
+     --path:
+        Path to directory for writing files
+     --num_threads:
+        Number of CPU cores to use for running simulations in parallel
+     --convert:
+        Convert files between Cantera and Chemkin formats (.cti <=> .inp)
+     --thermo:
+        thermodynamic data filename (only necessary for Chemkin files)
+     --transport:
+        transport data filename (only necessary for Chemkin files)
+
+Example reduction
+=================
+
+To apply the DRGEP method for reducing the GRI Mech 3.0 model (included with
+Cantera), with a maximum error of 10%, using the main reactants as targets
+(CH\ :sub:`4` and O\ :sub:`2`\ ) and always retaining N\ :sub:`2`\ ,
+execute this command:
+
+.. code-block:: bash
+
+    pymars -m gri30.cti --method DRGEP --error 10 \
+      --conditions input.yaml --targets CH4 O2 --retained_species N2
+
+Note that you need to create a file ``input.yaml`` with a list of autoignition
+initial conditions for the reduction; download an
+:download:`example file <example_input_file.yaml>` or see the
+:ref:`autoignition` section for a detailed explanation.
+
+**Parallelization:** If you have a large number of initial conditions, the
+reduction may be sped up by parallelizing the associated simulations over
+multiple CPU cores. You can do this by adding ``--num_threads N``, where ``N``
+is the desired number of cores. If you just specify ``--num_threads``, then
+pyMARS will use the available numbera of cores minus one.
+
+(pyMARS does not currently support distributed memory parallelization, meaning
+across multiple nodes that do not share the same memory.)
+
+**Sensitivity analysis:** To perform sensitivity analysis following DRGEP,
+add the option ``--sensitivity_analysis``. Select the type of sensitivity analysis
+with the ``--sensitivity_method`` option. Generally it is a good idea to specify an upper
+threshold value of 0.2 to ensure important species are not evaluated; do this
+by adding the option ``--upper_threshold 0.2``. (In the :ref:`sec_theory`
+guide, this variable is referred to as :math:`\epsilon^*`.)
+
+Then, the full command for performing a reduction using DRGEPSA with
+parallelized simalations would be:
+
+.. code-block:: bash
+
+   pymars -m gri30.cti --method DRGEP --error 10 \
+      --conditions input.yaml --targets CH4 O2 --retained_species N2 \
+      --num_threads --sensitivity_analysis --sensitvity_method initial --upper_threshold 0.2
+
+.. _autoignition:
+
+Autoignition conditions
+=======================
+
[…]
