### 2019-07-08 #1543 @kyleniemeyer
Hi @katyhuff, I just noticed that @jcsutherland  doesn't have a checklist and isn't listed as a reviewer above—I think that's because you had used the `@whedon assign ... as reviewer` rather than `add` for @fcontino. I think it's possible to add him manually at this point, not sure about adding via whedon.
