### 2019-07-15 #1543 @jcsutherland
@katyhuff  I don't seem to have access to check any boxes for the review.  When I follow the [invitation URL you supplied above](https://github.com/openjournals/joss-reviews/invitations), I see the following:

> Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account.
