### 2019-07-15 #1543 @kyleniemeyer
@arfon @danielskatz @katyhuff I just added @jcsutherland as a collaborator in the `joss-reviews` repo, which (I believe) is what whedon does when adding a reviewer at the pre-review stage.
