### 2019-07-15 Niemeyer-Research-Group/pyMARS#56 @jcsutherland
Specify appropriate units for input file in documentation
Documentation should include units for the following elements of the input file:
- [x] `pressure`
- [x] `temperature`
- [x] Mole/mass fractions for composition specification.  It might be nice to allow the user to specify either and convert internally
- [x] `end-time`

The sample input file should also include units in comments for clarity
