### 2019-07-15 Niemeyer-Research-Group/pyMARS#57 @jcsutherland
Consider consolidating required inputs to the yaml file
Presently, there are two places that required inputs reside: the command line and the input file.  This seems a bit fragmented.

Consider consolidating the _required_ inputs into the yaml file.  These include:

- `--model`
- `--targets` and `retained_speces`
- `--method`
- `--error`

This would improve reproducibility since all of the required inputs would be contained in a single input file.  Then consider renaming `--conditions` or make the input file a required non-named argument.
