### 2019-07-15 Niemeyer-Research-Group/pyMARS#58 @jcsutherland
Add error trapping for species inputs
If a species is "retained" or "target" and isn't found in the cantera input file, that error should be trapped.  It is presently silently ignored, which can give unexpected results.  Also note that the species names are case-sensitive.
