### 2019-07-15 Niemeyer-Research-Group/pyMARS#59 @jcsutherland
Don't assume that CH4 is present in the mechanism.
It appears that CH4 is assumed to be part of the mechanism somewhere?

# Error description:
When running against a mechanism that does not have `CH4` as a species, an exception occurs.
```
  File "interfaces/cython/cantera/thermo.pyx", line 603, in cantera._cantera.ThermoPhase.set_equivalence_ratio
  File "interfaces/cython/cantera/thermo.pyx", line 999, in cantera._cantera.ThermoPhase.TPX.__set__
  File "interfaces/cython/cantera/thermo.pyx", line 561, in cantera._cantera.ThermoPhase.X.__set__
cantera._cantera.CanteraError:
***********************************************************************
CanteraError thrown by Phase::setMoleFractionsByName:
Unknown species 'CH4'
***********************************************************************
```

# Steps to Reproduce
Unzip the attached file and then run the following:
```
pymars \
    --targets CO O2 H2 \
    --retained_species N2 CO CO2 H2O \
    --conditions coh2-conditions.yaml \
    --method DRGEP \
    --error 10 \
    -m coh2-yetter.cti
```
The `cti` and `conditions` files are contained in the attached zip file.
[coh2files.zip](https://github.com/Niemeyer-Research-Group/pyMARS/files/3393617/coh2files.zip)
