### 2019-07-15 Niemeyer-Research-Group/pyMARS#60 @jcsutherland
Cantera group names are hard-coded?
# Background

Cantera allows specification of various `ideal_gas` objects to include different species/elements/reactions from the input file.  It doesn't appear that `pyMARS` supports this, and it isn't clear if it is looking for a magic string somewhere.

```
ideal_gas(name = "gri30",
      elements = " O  H  C  N  Ar ",
      species = """ H2  H  O  O2  OH  H2O  HO2  H2O2  C  CH
                   CH2  CH2(S)  CH3  CH4  CO  CO2  HCO  CH2O  CH2OH  CH3O
                   CH3OH  C2H  C2H2  C2H3  C2H4  C2H5  C2H6  HCCO  CH2CO  HCCOH
                   N  NH  NH2  NH3  NNH  NO  NO2  N2O  HNO  CN
                   HCN  H2CN  HCNN  HCNO  HOCN  HNCO  NCO  N2  AR  C3H7
                   C3H8  CH2CHO  CH3CHO """,
      reactions = "all",
      kinetics = "GRI30",
      initial_state = state(temperature = 300.0,
                        pressure = OneAtm)    )

ideal_gas(name = "gri30_mix",
      elements = " O  H  C  N  Ar ",
      species = """ H2  H  O  O2  OH  H2O  HO2  H2O2  C  CH
                   CH2  CH2(S)  CH3  CH4  CO  CO2  HCO  CH2O  CH2OH  CH3O
                   CH3OH  C2H  C2H2  C2H3  C2H4  C2H5  C2H6  HCCO  CH2CO  HCCOH
                   N  NH  NH2  NH3  NNH  NO  NO2  N2O  HNO  CN
                   HCN  H2CN  HCNN  HCNO  HOCN  HNCO  NCO  N2  AR  C3H7
                   C3H8  CH2CHO  CH3CHO """,
      reactions = "all",
      kinetics = "GRI30",
      transport = "Mix",
      initial_state = state(temperature = 300.0,
                        pressure = OneAtm)    )

ideal_gas(name = "gri30_multi",
      elements = " O  H  C  N  Ar ",
      species = """ H2  H  O  O2  OH  H2O  HO2  H2O2  C  CH
                   CH2  CH2(S)  CH3  CH4  CO  CO2  HCO  CH2O  CH2OH  CH3O
                   CH3OH  C2H  C2H2  C2H3  C2H4  C2H5  C2H6  HCCO  CH2CO  HCCOH
                   N  NH  NH2  NH3  NNH  NO  NO2  N2O  HNO  CN
                   HCN  H2CN  HCNN  HCNO  HOCN  HNCO  NCO  N2  AR  C3H7
                   C3H8  CH2CHO  CH3CHO """,
      reactions = "all",
      kinetics = "GRI30",
      transport = "Multi",
      initial_state = state(temperature = 300.0,
                        pressure = OneAtm)    )
```
It appears that this does not play well with `pyMARS`

# Error description
When running against a `cti` file including the above examples, the following exception occurs:
```
  File "interfaces/cython/cantera/base.pyx", line 29, in cantera._cantera._SolutionBase.__cinit__
  File "interfaces/cython/cantera/base.pyx", line 78, in cantera._cantera._SolutionBase._init_cti_xml
cantera._cantera.CanteraError:
***********************************************************************
CanteraError thrown by Factory::create:
No such type: 'gri30'
***********************************************************************
```
# Steps to reproduce
Unzip the attached file and then run the following:
```
#!/bin/sh
pymars \
    --targets CH4 O2 H2 \
    --retained_species N2 CO CO2 H2O \
    --conditions methane-conditions.yaml \
    --method DRGEP \
    --error 10 \
    -m methane-gri30.cti
```
[methanefiles.zip](https://github.com/Niemeyer-Research-Group/pyMARS/files/3393668/methanefiles.zip)
