### 2019-07-15 Niemeyer-Research-Group/pyMARS#61 @jcsutherland
Default to steady-state solutions rather than specifying `end_time`?
I find it a little strange that the default behavior isn't to solve the ignition problem to steady state.  Specifying `end_time` seems like something that a more advanced user may be interested in, but typically we would want a steady solution, wouldn't we?
