### 2019-07-15 #1543 @jcsutherland
# jcsutherland  Review Comments

## Critical issues to be resolved prior to acceptance
- [ ]  **Unit tests** should be implemented.  See [issue 26](https://github.com/Niemeyer-Research-Group/pyMARS/issues/26).  It appears that this is in progress, but the documentation doesn't indicate how to run tests.
- [ ] The [documentation](https://niemeyer-research-group.github.io/pyMARS/usage.html) should clearly specify _**units**_ expected in the input file.  See [issue 56](https://github.com/Niemeyer-Research-Group/pyMARS/issues/56)
- [ ]  Raise errors if a species from the inputs isn't found in the Cantera file.  See [issue 58](https://github.com/Niemeyer-Research-Group/pyMARS/issues/58)
- [ ] pyMARS presently has a few critical bugs that should be corrected prior to acceptance:
   - [ ] CH4 is assumed to be present in the mechanism.  See [issue 59](https://github.com/Niemeyer-Research-Group/pyMARS/issues/59)
  - [ ] Specifying Cantera groups causes exceptions.  See [issue 60](https://github.com/Niemeyer-Research-Group/pyMARS/issues/60)

## Other issues to consider
- [ ] Consider required consolidating inputs to the input file.  See [issue 57](https://github.com/Niemeyer-Research-Group/pyMARS/issues/57).  This isn't a showstopper for me, just something for the developers to consider since I think that it would improve user workflow.
- [ ] Consider making a stead-state solve the default behavior.  See [issue 61](https://github.com/Niemeyer-Research-Group/pyMARS/issues/61).  I don't think that this is a major issue - more food for thought.
