### 2019-07-15 #1543 @kyleniemeyer
Hi @jcsutherland, thanks so much for all your feedback! We will start to address these items, but quickly, pyMARS does actually already have an automated testing suite, based on pytest.

If you have pytest installed, navigating to the repo and running `pytest` should automatically find and run all the tests, contained in the `pymars/tests` directory.

However, I do see that this isn't described in the documentation, so we will add that.
