### 2019-07-15 Niemeyer-Research-Group/pyMARS#59 @kyleniemeyer
@jcsutherland it looks like your `coh2-conditions.yaml` file with autoignition initial conditions has `CH4` listed as the fuel in the first two items:
```
  fuel:
      CH4: 1.0
```
so the code raises an error when attempting to specify the initial mole fractions.

However, the `check_inputs()` function should really raise an error message for this, so that it can be better diagnosed.
