### 2019-07-15 Niemeyer-Research-Group/pyMARS#59 @jcsutherland
Hehe - you're right.  That was a copy/paste error on my part.
