### 2019-07-15 Niemeyer-Research-Group/pyMARS#58 @kyleniemeyer
Good catch, thanks.
