### 2019-07-15 Niemeyer-Research-Group/pyMARS#59 @kyleniemeyer
@jcsutherland no problem. I'm still going to leave this issue open, since such an issue should be caught by the code.
