### 2019-07-15 Niemeyer-Research-Group/pyMARS#56 @kyleniemeyer
The docs do reference the units in this section https://niemeyer-research-group.github.io/pyMARS/usage.html#autoignition-conditions, but I agree we can list these out explicitly for more clarity (and include in the example).
