### 2019-07-15 Niemeyer-Research-Group/pyMARS#61 @kyleniemeyer
Good point. `end_time` isn't required, but the default time used internally is 10 sec, which isn't great either.

It shouldn't be too difficult to have it integrate to steady state, although when thinking about this before I was a little concerned about it getting stuck integrating forever... not sure if I can think of when that would actually happen, though.
