### 2019-07-15 Niemeyer-Research-Group/pyMARS#57 @kyleniemeyer
I struggled with the best way to handle this, but I think I agree that reduction inputs make sense in a YAML file, also including `--sensitivity_analysis`, `--sensitivity_type`, and `--upper_threshold`.

The remaining command-line options would be:
```
-h, --help:
   show this help message and exit
--path:
   Path to directory for writing files
--num_threads:
   Number of CPU cores to use for running simulations in parallel
--convert:
   Convert files between Cantera and Chemkin formats (.cti <=> .inp)
--thermo:
   thermodynamic data filename (only necessary for Chemkin files)
--transport:
   transport data filename (only necessary for Chemkin files)
```

with a new `--input` option for the reduction input YAML file, that includes the autoignition conditions list.

(`--thermo` and `--transport` make sense to me  as command-line inputs, since you would only need to use these once before you have a Cantera-format model to use.)
