### 2019-07-15 Niemeyer-Research-Group/pyMARS#60 @kyleniemeyer
Hmm, this one is a little tricky. Right now pyMARS assumes a single `ideal_gas` object, but you're right that some files will have multiple.

What it *should* do is modify the lists of species and reactions in each object.

Where things get tricky is if the `ideal_gas` objects include different sets of species and/or reactions. I'll need to look into this a bit more.
