+++ pymars/sampling.py
@@ -2,7 +2,7 @@
 import os
 import multiprocessing
 import logging
-from typing import NamedTuple
+from typing import NamedTuple, Dict
 
 import numpy as np
 import yaml
@@ -10,20 +10,35 @@ import cantera as ct
 
 from .simulation import Simulation
 
-class SamplingInputs(NamedTuple):
-    """Collects input and output files for sampling (autoignition, PSR, laminar flame)
+data_files = {
+    'data_ignition': 'ignition_data.dat', 'output_ignition': 'ignition_output.txt',
+    'data_psr': 'psr_data.dat', 'output_psr': 'psr_output.txt',
+    'data_flame': 'laminarflame_data.dat', 'output_flame': 'laminarflame_output.txt'
+    }
+
+class InputIgnition(NamedTuple):
+    """Holds input parameters for a single autoignition case.
     """
-    input_ignition: str = ''
-    data_ignition: str = 'ignition_data.dat'
-    output_ignition: str = 'ignition_output.txt'
+    kind: str
+    temperature: float
+    pressure: float
 
-    input_psr: str = ''
-    data_psr: str = 'psr_data.dat'
-    output_psr: str = 'psr_output.txt'
+    end_time: float = 0.0
 
-    input_laminar_flame: str = ''
-    data_laminar_flame: str = 'laminarflame_data.dat'
-    output_laminar_flame: str = 'laminarflame_output.txt'
+    equivalence_ratio: float = 0.0
+    fuel: Dict = {}
+    oxidizer: Dict = {}
+    reactants: Dict = {}
+
+
+class InputPSR(NamedTuple):
+    """Holds input parameters for single PSR simulation.
+    """
+
+
+class InputLaminarFlame(NamedTuple):
+    """Holds input parameters for single laminar flame simulation.
+    """
 
 
 def simulation_worker(sim_tuple):
@@ -94,13 +109,17 @@ def calculate_error(metrics_original, metrics_test):
     return error
 
 
-def read_metrics(inputs):
+def read_metrics(ignition_conditions, psr_conditions=[], flame_conditions=[]):
     """Reads in stored already-sampled metrics.
 
     Parameters
     ----------
-    inputs : SamplingInputs
-        Inputs necessary for sampling
+    ignition_conditions : list of InputIgnition
+        List of autoignition initial conditions.
+    psr_conditions : list of InputPSR, optional
+        List of PSR simulation conditions.
+    flame_conditions : list of InputLaminarFlame, optional
+        List of laminar flame simulation conditions.
 
     Returns
     -------
@@ -108,34 +127,39 @@ def read_metrics(inputs):
         Calculated metrics for model, used for evaluating error
 
     """
-    if inputs.input_ignition:
-        exists_output = os.path.isfile(inputs.output_ignition)
+    if ignition_conditions:
+        exists_output = os.path.isfile(data_files['output_ignition'])
         if exists_output:
-            ignition_delays = np.genfromtxt(inputs.output_ignition, delimiter=',')
+            ignition_delays = np.genfromtxt(data_files['output_ignition'], delimiter=',')
         else:
             raise SystemError('Error, no ignition output file present.')
 
-    elif inputs.input_psr:
+    if psr_conditions:
         raise NotImplementedError('PSR calculations not currently supported.')
-    elif inputs.input_laminar_flame:
+    
+    if flame_conditions:
         raise NotImplementedError('Laminar flame calculations not currently supported.')
-    else:
-        raise KeyError('No input files specified for sampling.')
 
     return ignition_delays
 
 
-def sample_metrics(inputs, model, num_threads=1, path='', reuse_saved=False):
+def sample_metrics(model, ignition_conditions, psr_conditions=[], flame_conditions=[],
[…]
