### 2019-07-19 c1efc08
moved all test files to assets directory for cleanup
