+++ pymars/tests/test_drgep.py
@@ -659,8 +659,12 @@ class TestRunDRGEP:
                 fuel={'CH4': 1.0}, oxidizer={'O2': 1.0, 'N2': 3.76}
                 ),
         ]
-        data_files['output_ignition'] = relative_location('example_ignition_output.txt')
-        data_files['data_ignition'] = relative_location('example_ignition_data.dat')
+        data_files['output_ignition'] = relative_location(
+            os.path.join('assets', 'example_ignition_output.txt')
+            )
+        data_files['data_ignition'] = relative_location(
+            os.path.join('assets', 'example_ignition_data.dat')
+            )
         error = 5.0
 
         # Run DRG
@@ -671,7 +675,7 @@ class TestRunDRGEP:
                 )
 
         # Expected answer
-        expected_model = ct.Solution(relative_location('drgep_gri30.cti'))
+        expected_model = ct.Solution(relative_location(os.path.join('assets', 'drgep_gri30.cti')))
         
         # Make sure models are the same
         assert check_equal(reduced_model.model.species_names, expected_model.species_names)
