+++ pymars/tests/test_drg.py
@@ -525,7 +525,10 @@ class TestReduceDRG:
                 ),
         ]
         
-        data = np.genfromtxt(relative_location('example_ignition_data.dat'), delimiter=',')
+        data = np.genfromtxt(
+            relative_location(os.path.join('assets', 'example_ignition_data.dat')), 
+            delimiter=','
+            )
 
         model = ct.Solution(model_file)
         matrices = []
@@ -566,8 +569,12 @@ class TestRunDRG:
                 fuel={'CH4': 1.0}, oxidizer={'O2': 1.0, 'N2': 3.76}
                 ),
         ]
-        data_files['output_ignition'] = relative_location('example_ignition_output.txt')
-        data_files['data_ignition'] = relative_location('example_ignition_data.dat')
+        data_files['output_ignition'] = relative_location(
+            os.path.join('assets', 'example_ignition_output.txt')
+            )
+        data_files['data_ignition'] = relative_location(
+            os.path.join('assets', 'example_ignition_data.dat')
+            )
         error = 5.0
 
         # Run DRG
@@ -578,7 +585,7 @@ class TestRunDRG:
                 )
 
         # Expected answer
-        expected_model = ct.Solution(relative_location("drg_gri30.cti"))
+        expected_model = ct.Solution(relative_location(os.path.join('assets', 'drg_gri30.cti')))
         
         # Make sure models are the same
         assert check_equal(reduced_model.model.species_names, expected_model.species_names)
