### 2019-07-19 d34318b
Updates docs for new input file. Closes #56
