+++ docs/usage.rst
@@ -48,23 +48,56 @@ The following options are available, and can also be seen by using the
      --transport:
         transport data filename (only necessary for Chemkin files)
 
+
 Example reduction
 =================
 
 To apply the DRGEP method for reducing the GRI Mech 3.0 model (included with
 Cantera), with a maximum error of 10%, using the main reactants as targets
 (CH\ :sub:`4` and O\ :sub:`2`\ ) and always retaining N\ :sub:`2`\ ,
-execute this command:
+first create a YAML file (let's call it ``reduction_input.yaml``:
+
+.. code-block:: yaml
+
+    model: gri30.cti
+    targets:
+      - CH4
+      - O2
+    retained-species:
+      - N2
+    method: DRGEP
+    error: 10.0
+    sensitivity-analysis: False
+    autoignition-conditions:
+      - kind: constant volume
+        pressure: 1.0
+        temperature: 1000.0
+        fuel:
+          CH4: 1.0
+        oxidizer:
+          O2: 1.0
+          N2: 3.76
+        equivalence-ratio: 1.0
+
+      - kind: constant volume
+        pressure: 1.0
+        temperature: 1200.0
+        fuel:
+          CH4: 1.0
+        oxidizer:
+          O2: 1.0
+          N2: 3.76
+        equivalence-ratio: 0.5
+
+Then, execute this command:
 
 .. code-block:: bash
 
-    pymars -m gri30.cti --method DRGEP --error 10 \
-      --conditions input.yaml --targets CH4 O2 --retained_species N2
+    pymars --input reduction_input.yaml
 
-Note that you need to create a file ``input.yaml`` with a list of autoignition
-initial conditions for the reduction; download an
-:download:`example file <example_input_file.yaml>` or see the
-:ref:`autoignition` section for a detailed explanation.
+You can also download an annotated version of the
+:download:`example input file <example_input_file.yaml>`; refer to the
+:ref:`inputfile` section for a detailed explanation.
 
 **Parallelization:** If you have a large number of initial conditions, the
 reduction may be sped up by parallelizing the associated simulations over
@@ -76,85 +109,134 @@ pyMARS will use the available numbera of cores minus one.
 across multiple nodes that do not share the same memory.)
 
 **Sensitivity analysis:** To perform sensitivity analysis following DRGEP,
-add the option ``--sensitivity_analysis``. Select the type of sensitivity analysis
-with the ``--sensitivity_method`` option. Generally it is a good idea to specify an upper
+change the ``sensitivity-analysis`` key to ``True`` in the input file,
+and choose the type of sensitivity analysis with the ``sensitivity-type`` field
+(either ``initial`` or ``greedy``). Let's use the "initial" method for now, since
+it is less computationally expensive.
+
+Generally it is a good idea to specify an upper
 threshold value of 0.2 to ensure important species are not evaluated; do this
-by adding the option ``--upper_threshold 0.2``. (In the :ref:`sec_theory`
+by adding the line ``upper-threshold: 0.2``. (In the :ref:`sec_theory`
 guide, this variable is referred to as :math:`\epsilon^*`.)
 
-Then, the full command for performing a reduction using DRGEPSA with
+Your new input file (called ``drgepsa_input.yaml``) would then look like:
+
+.. code-block:: yaml
+
+    model: gri30.cti
+    targets:
+      - CH4
+      - O2
+    retained-species:
+      - N2
+    method: DRGEP
+    error: 10.0
+    sensitivity-analysis: True
+    sensitivity-type: initial
+    upper-threshold: 0.2
+    autoignition-conditions:
+      - kind: constant volume
+        pressure: 1.0
+        temperature: 1000.0
+        fuel:
+          CH4: 1.0
+        oxidizer:
+          O2: 1.0
+          N2: 3.76
[…]
