### 2019-07-19 a6aa4fa
Adds way to specify reactants mass fraction; addresses #56
