### 2019-07-20 Niemeyer-Research-Group/pyMARS#60 @kyleniemeyer
@jcsutherland so, this seems to be an issue with the `methane-gri30.cti` file you are using; if I just try to create a `cantera.Solution` object with it, Cantera throws the error above:

```
gas = ct.Solution('methane-gri30.cti')

CanteraError                              Traceback (most recent call last)
<ipython-input-2-3045edfef451> in <module>
----> 1 gas = ct.Solution('methane-gri30.cti')

interfaces/cython/cantera/base.pyx in cantera._cantera._SolutionBase.__cinit__()

interfaces/cython/cantera/base.pyx in cantera._cantera._SolutionBase._init_cti_xml()

CanteraError:
***********************************************************************
CanteraError thrown by Factory::create:
No such type: 'gri30'
***********************************************************************
```

I think the issue is the `kinetics = "GRI30",` line that appears in each `ideal_gas` object. If I remove that, the file loads without error; in fact, the [GRI Mech 3.0 included with Cantera](https://github.com/Cantera/cantera/blob/master/data/inputs/gri30.cti) does not have that line, and this is the model we use for a bunch of unit tests.

However, regarding the presence of multiple `ideal_gas` objects: the current behavior of pyMARS is to reduce whichever is loaded by default when creating a `Cantera.Solution` object from the CTI file. In the case of `gri30.cti`, this is the first one, without the transport field specified. The retained `species` objects contain their transport data, though.
