### 2019-07-20 Niemeyer-Research-Group/pyMARS#60 @kyleniemeyer
In the situation that the CTI file contains multiple `ideal_gas` objects, I don't think we want to try reducing each of them (in case they contain varying numbers/sets of species or reactions).

Instead, I think the best solution is to include the name of the object to be reduced as an optional reduction input (now contained in a single YAML file, per e6ec139).
