### 2019-07-21 0ba0b96
Adds optional input parameter for phase name. Closes #60
Previously, pyMARS reduced the ideal_gas() object that Cantera
loads by default from a CTI file, but there can be multiple objects.
The `phase-name` key in the input file allows a user to specify
a specific ideal_gas() object for reduction. This required adding
and passing an attribute through many functions.
