+++ pymars/drg.py
@@ -118,8 +118,8 @@ def trim_drg(matrix, species_names, species_targets, threshold):
 
 def reduce_drg(model_file, species_targets, species_safe, threshold, 
                matrices, ignition_conditions, sampled_metrics, 
-               previous_model=None, threshold_upper=None, num_threads=1,
-               path=''
+               phase_name='', previous_model=None, threshold_upper=None, 
+               num_threads=1, path=''
                ):
     """Given a threshold and DRG matrix, reduce the model and determine the error.
 
@@ -139,6 +139,8 @@ def reduce_drg(model_file, species_targets, species_safe, threshold,
         List of autoignition initial conditions.
     sampled_metrics: numpy.ndarray
         Global metrics from original model used to evaluate error
+    phase_name : str, optional
+        Optional name for phase to load from CTI file (e.g., 'gas'). 
     previous_model : ReducedModel, optional
         Model produced at previous threshold level; used to avoid repeated work.
     threshold_upper : float, optional
@@ -158,7 +160,7 @@ def reduce_drg(model_file, species_targets, species_safe, threshold,
         Return reduced model and associated metadata
 
     """
-    solution = ct.Solution(model_file)
+    solution = ct.Solution(model_file, phase_name)
 
     species_retained = []
     for matrix in matrices:
@@ -175,13 +177,16 @@ def reduce_drg(model_file, species_targets, species_safe, threshold,
                        ]
 
     # Cut the exclusion list from the model.
-    reduced_model = trim(model_file, species_removed, f'reduced_{model_file}')
+    reduced_model = trim(
+        model_file, species_removed, f'reduced_{model_file}', phase_name=phase_name
+        )
     reduced_model_filename = soln2cti.write(
         reduced_model, f'reduced_{reduced_model.n_species}.cti', path=path
         )
 
     reduced_model_metrics = sample_metrics(
-        reduced_model_filename, ignition_conditions, num_threads=num_threads, path=path
+        reduced_model_filename, ignition_conditions, phase_name=phase_name, 
+        num_threads=num_threads, path=path
         )
     error = calculate_error(sampled_metrics, reduced_model_metrics)
     
@@ -203,7 +208,7 @@ def reduce_drg(model_file, species_targets, species_safe, threshold,
 
 
 def run_drg(model_file, ignition_conditions, psr_conditions, flame_conditions, 
-            error_limit, species_targets, species_safe, 
+            error_limit, species_targets, species_safe, phase_name='',
             threshold_upper=None, num_threads=1, path=''
             ):
     """Main function for running DRG reduction.
@@ -224,6 +229,8 @@ def run_drg(model_file, ignition_conditions, psr_conditions, flame_conditions,
         List of target species names
     species_safe : list of str
         List of species names to always be retained
+    phase_name : str, optional
+        Optional name for phase to load from CTI file (e.g., 'gas'). 
     threshold_upper: float, optional
         Upper threshold (epsilon^*) to identify limbo species for sensitivity analysis
     num_threads : int, optional
@@ -240,7 +247,7 @@ def run_drg(model_file, ignition_conditions, psr_conditions, flame_conditions,
         Return reduced model and associated metadata
 
     """
-    solution = ct.Solution(model_file)
+    solution = ct.Solution(model_file, phase_name)
 
     assert species_targets, 'Need to specify at least one target species.'
 
@@ -248,7 +255,7 @@ def run_drg(model_file, ignition_conditions, psr_conditions, flame_conditions,
     # (e.g, ignition delays). Also produce adjacency matrices for graphs, which
     # will be used to produce graphs for any threshold value.
     sampled_metrics, sampled_data = sample(
[…]

[…]
