### 2019-07-22 daf2800
Changes integration behavior to reach steady state by default. Closes #61.
By default, autoignition simulations now integrate to steady state
(stolen from Cantera's advance_to_steady_state method), with a maximum
10000 steps to protect against integrating indefinitely. Users can
specify `end-time` for a specific end time or change the maximum
number of steps with `max-steps` in the associated input block,
to override this behavior.
