### 2019-07-22 6cf3f21
trying to fix Azure Pipelines build
