+++ azure-pipelines.yml
@@ -16,6 +16,7 @@ jobs:
     displayName: 'Add conda to PATH'
   
   - script: |
+      call C:\Miniconda3\Scripts\activate
       conda config --set always_yes yes --set changeps1 no
       conda config --append channels conda-forge
       conda update conda
@@ -27,6 +28,7 @@ jobs:
 
   - script: |
       conda env create --quiet --file test-environment.yaml
+      call activate py$(python.version)
       conda info -a
     displayName: 'Create Anaconda environment'
   
@@ -37,7 +39,6 @@ jobs:
     displayName: 'List package in environment'
 
   - script: |
-      call activate py$(python.version)
       pip install pytest-azurepipelines
       pytest -v --cov=./ --cov-report html
     displayName: 'test with pytest'
