### 2019-07-22 060ca6e
More attempted fixes for Pipelines
