### 2019-07-22 5364abd
More attempted fixes for Pipelines
