### 2019-07-23 2e14cdc
having Azure build install pytest-cov [skip travis]
