+++ azure-pipelines.yml
@@ -39,6 +39,6 @@ jobs:
     displayName: 'List package in environment'
 
   - script: |
-      pip install pytest-azurepipelines
+      pip install pytest-cov pytest-azurepipelines
       pytest -v --cov=./ --cov-report html
     displayName: 'test with pytest'
