### 2019-07-23 f7326e8
[skip travis] adding additional conda activate command in Azure build
