+++ azure-pipelines.yml
@@ -39,6 +39,7 @@ jobs:
     displayName: 'List package in environment'
 
   - script: |
+      call activate py$(python.version)
       pip install pytest-cov pytest-azurepipelines
       pytest -v --cov=./ --cov-report html
     displayName: 'test with pytest'
