### 2019-07-23 524c636
Bump version v1.0.0a10
