+++ _sources/usage.rst.txt
@@ -15,32 +15,17 @@ where your input files are located, then run::
 The following options are available, and can also be seen by using the
 ``-h`` or ``--help`` option:
 
-.. code-block:: bash
+.. code-block:: none
 
      -h, --help:
         show this help message and exit
-     -m, --model:
-        input model filename (e.g., mech.cti)
-     -e, --error:
-        Maximum error percentage for the reduced model
-     --method {DRG,DRGEP,PFA}
-        skeletal reduction method to use
-     --conditions:
-        File with list of autoignition initial conditions
-     --targets:
-        List of target species (e.g., "CH4 O2")
-     --retained_species:
-        List of non-target species to always retain (e.g., "N2 Ar")
-     --sensitivity_analysis:
-        Run sensitivity analysis after completing another method
-     --sensitivity_type {initial, greedy}
-        Sensitivity analysis method to use
-     --upper_threshold:
-        Upper threshold value used to determine species for sensitivity analysis
+     -i, --input:
+        YAML file with reduction inputs
      --path:
         Path to directory for writing files
      --num_threads:
-        Number of CPU cores to use for running simulations in parallel
+        Number of CPU cores to use for running simulations in parallel.
+        If no number, then use available number of cores minus 1.
      --convert:
         Convert files between Cantera and Chemkin formats (.cti <=> .inp)
      --thermo:
@@ -48,23 +33,56 @@ The following options are available, and can also be seen by using the
      --transport:
         transport data filename (only necessary for Chemkin files)
 
+
 Example reduction
 =================
 
 To apply the DRGEP method for reducing the GRI Mech 3.0 model (included with
 Cantera), with a maximum error of 10%, using the main reactants as targets
 (CH\ :sub:`4` and O\ :sub:`2`\ ) and always retaining N\ :sub:`2`\ ,
-execute this command:
+first create a YAML file (let's call it ``reduction_input.yaml``:
+
+.. code-block:: yaml
+
+    model: gri30.cti
+    targets:
+      - CH4
+      - O2
+    retained-species:
+      - N2
+    method: DRGEP
+    error: 10.0
+    sensitivity-analysis: False
+    autoignition-conditions:
+      - kind: constant volume
+        pressure: 1.0
+        temperature: 1000.0
+        fuel:
+          CH4: 1.0
+        oxidizer:
+          O2: 1.0
+          N2: 3.76
+        equivalence-ratio: 1.0
+
+      - kind: constant volume
+        pressure: 1.0
+        temperature: 1200.0
+        fuel:
+          CH4: 1.0
+        oxidizer:
+          O2: 1.0
+          N2: 3.76
+        equivalence-ratio: 0.5
+
+Then, execute this command:
 
 .. code-block:: bash
 
-    pymars -m gri30.cti --method DRGEP --error 10 \
-      --conditions input.yaml --targets CH4 O2 --retained_species N2
+    pymars --input reduction_input.yaml
 
-Note that you need to create a file ``input.yaml`` with a list of autoignition
-initial conditions for the reduction; download an
-:download:`example file <example_input_file.yaml>` or see the
-:ref:`autoignition` section for a detailed explanation.
+You can also download an annotated version of the
+:download:`example input file <example_input_file.yaml>`; refer to the
+:ref:`inputfile` section for a detailed explanation.
 
 **Parallelization:** If you have a large number of initial conditions, the
 reduction may be sped up by parallelizing the associated simulations over
@@ -76,85 +94,155 @@ pyMARS will use the available numbera of cores minus one.
 across multiple nodes that do not share the same memory.)
 
 **Sensitivity analysis:** To perform sensitivity analysis following DRGEP,
[…]
