### 2019-07-23 2c3ca7e
Update docs after building Travis build 98 of Niemeyer-Research-Group/pyMARS
The docs were built from the tag 'v1.0.0a10' against the commit
524c636329f98fe26e7e56ee635e762f6bc877d1.

The Travis build that generated this commit is at
https://travis-ci.org/Niemeyer-Research-Group/pyMARS/jobs/562393135.

The doctr command that was run is

    /home/travis/miniconda/bin/doctr deploy --build-tags tags/v1.0.0a10
