+++ tags/v1.0.0a10/_sources/usage.rst.txt
@@ -1,0 +1,272 @@
+.. _sec_usage:
+
+=====
+Usage
+=====
+
+.. toctree::
+    :maxdepth: 1
+
+Once installed, pyMARS can be run as an executable. Change to the directory
+where your input files are located, then run::
+
+    pymars [options]
+
+The following options are available, and can also be seen by using the
+``-h`` or ``--help`` option:
+
+.. code-block:: none
+
+     -h, --help:
+        show this help message and exit
+     -i, --input:
+        YAML file with reduction inputs
+     --path:
+        Path to directory for writing files
+     --num_threads:
+        Number of CPU cores to use for running simulations in parallel.
+        If no number, then use available number of cores minus 1.
+     --convert:
+        Convert files between Cantera and Chemkin formats (.cti <=> .inp)
+     --thermo:
+        thermodynamic data filename (only necessary for Chemkin files)
+     --transport:
+        transport data filename (only necessary for Chemkin files)
+
+
+Example reduction
+=================
+
+To apply the DRGEP method for reducing the GRI Mech 3.0 model (included with
+Cantera), with a maximum error of 10%, using the main reactants as targets
+(CH\ :sub:`4` and O\ :sub:`2`\ ) and always retaining N\ :sub:`2`\ ,
+first create a YAML file (let's call it ``reduction_input.yaml``:
+
+.. code-block:: yaml
+
+    model: gri30.cti
+    targets:
+      - CH4
+      - O2
+    retained-species:
+      - N2
+    method: DRGEP
+    error: 10.0
+    sensitivity-analysis: False
+    autoignition-conditions:
+      - kind: constant volume
+        pressure: 1.0
+        temperature: 1000.0
+        fuel:
+          CH4: 1.0
+        oxidizer:
+          O2: 1.0
+          N2: 3.76
+        equivalence-ratio: 1.0
+
+      - kind: constant volume
+        pressure: 1.0
+        temperature: 1200.0
+        fuel:
+          CH4: 1.0
+        oxidizer:
+          O2: 1.0
+          N2: 3.76
+        equivalence-ratio: 0.5
+
+Then, execute this command:
+
+.. code-block:: bash
+
+    pymars --input reduction_input.yaml
+
+You can also download an annotated version of the
+:download:`example input file <example_input_file.yaml>`; refer to the
+:ref:`inputfile` section for a detailed explanation.
+
+**Parallelization:** If you have a large number of initial conditions, the
+reduction may be sped up by parallelizing the associated simulations over
+multiple CPU cores. You can do this by adding ``--num_threads N``, where ``N``
+is the desired number of cores. If you just specify ``--num_threads``, then
+pyMARS will use the available numbera of cores minus one.
+
+(pyMARS does not currently support distributed memory parallelization, meaning
+across multiple nodes that do not share the same memory.)
+
+**Sensitivity analysis:** To perform sensitivity analysis following DRGEP,
+change the ``sensitivity-analysis`` key to ``True`` in the input file,
+and choose the type of sensitivity analysis with the ``sensitivity-type`` field
+(either ``initial`` or ``greedy``). Let's use the "initial" method for now, since
+it is less computationally expensive.
+
+Generally it is a good idea to specify an upper
+threshold value of 0.2 to ensure important species are not evaluated; do this
+by adding the line ``upper-threshold: 0.2``. (In the :ref:`sec_theory`
+guide, this variable is referred to as :math:`\epsilon^*`.)
+
+Your new input file (called ``drgepsa_input.yaml``) would then look like:
+
+.. code-block:: yaml
+
+    model: gri30.cti
+    targets:
+      - CH4
+      - O2
+    retained-species:
+      - N2
+    method: DRGEP
+    error: 10.0
+    sensitivity-analysis: True
[…]
