### 2019-07-23 #1543 @jcsutherland
I'm good with acceptance at this point.
