### 2019-07-23 #1543 @katyhuff
Thanks @jcsutherland ! I appreciate your thorough review.

@fcontino: Thank you again for volunteering to review. I know you initially expressed some scheduling challenges. Will you be able to address this review sometime soon?  I hope you'll let us know what deadline will work for you.
