### 2019-07-23 #1543 @kyleniemeyer
Thanks @jcsutherland!

FYI @fcontino, we released v1.1.0a0 based on the changes made in response to @jcsutherland's comments, if you grabbed an earlier version.
