### 2019-07-23 20b33cb
Fixed checking for -m command-line input
