### 2019-07-23 2000c8c
Fixed additional reference to args.model
