### 2019-07-23 035f178
Removes erroneous warning message about integration not converging
