### 2019-07-24 f734de8
Fixes bug with multiple conditions in DRGEP
Importance coefficients were not being correctly calculated in DRGEP
with multiple conditions (sampled data), due to maximum dictionary values
not being stored correctly. These changes fix this, and add a test to check.
