### 2019-07-24 930f1b5
Bump v1.0.0a3
