### 2019-07-24 514a320
Trim pipelines build
