### 2019-07-24 1b9467f
Update docs after building Travis build 108 of Niemeyer-Research-Group/pyMARS
The docs were built from the tag 'v1.1.0a4' against the commit
514a32036193418a91fb5d5232d5fd440c832651.

The Travis build that generated this commit is at
https://travis-ci.org/Niemeyer-Research-Group/pyMARS/jobs/562868834.

The doctr command that was run is

    /home/travis/miniconda/bin/doctr deploy --build-tags tags/v1.1.0a4
