### 2019-07-24 36141b2
Fixed potential printing bug in reduced models with hyphens in names
