+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed bug in DRGEP overall interaction coefficient calculation with multiple conditions
 - Removed erroneous message from ignition simulations about lack of convergence
+- Fixed potenetial bug when printing CTI files with hyphens in species names
 
 
 ## [1.0.0] - 2019-06-21
