### 2019-07-24 4f2db59
Update docs after building Travis build 109 of Niemeyer-Research-Group/pyMARS
The docs were built from the tag 'v1.1.0a5' against the commit
625ba7c7a1effacff610e7ee8bb91c99aebec3ce.

The Travis build that generated this commit is at
https://travis-ci.org/Niemeyer-Research-Group/pyMARS/jobs/562920672.

The doctr command that was run is

    /home/travis/miniconda/bin/doctr deploy --build-tags tags/v1.1.0a5
