### 2019-07-25 e69e419
Bump 1.1.0a6; fixes bug printing efficiencies for falloff reactions
