### 2019-07-25 18c8bef
Update docs after building Travis build 111 of Niemeyer-Research-Group/pyMARS
The docs were built from the tag 'v1.1.0a6' against the commit
e69e4193fd23b611142339dc09bf6cd17cf7ffec.

The Travis build that generated this commit is at
https://travis-ci.org/Niemeyer-Research-Group/pyMARS/jobs/563737808.

The doctr command that was run is

    /home/travis/miniconda/bin/doctr deploy --build-tags tags/v1.1.0a6
