+++ tags/v1.1.0a6/_sources/reduce_model.rst.txt
@@ -1,0 +1,5 @@
+============
+reduce_model
+============
+
+.. automodule:: pymars.reduce_model
