### 2019-07-29 e111e95
Fixes bug that retained reactions where explicit third body removed.
Previously, trim() did not recognize that reactions with explicit
third bodies which have been removed should also be removed. This
commit fixes the bug, and also adds a test for this.
