### 2019-07-29 e8bfc5c
Update docs after building Travis build 116 of Niemeyer-Research-Group/pyMARS
The docs were built from the tag 'v1.1.0a8' against the commit
aace7aff2006788da1186f16e2d186ccbc3c1395.

The Travis build that generated this commit is at
https://travis-ci.org/Niemeyer-Research-Group/pyMARS/jobs/565174748.

The doctr command that was run is

    /home/travis/miniconda/bin/doctr deploy --build-tags tags/v1.1.0a8
