### 2019-08-07 #1543 @fcontino
I have installed pyMARS successfully on my Mac.

I am proceeding with some tests.

My first concern is that after installing, `pymars --help` does not give me the same as stated [here](https://niemeyer-research-group.github.io/pyMARS/usage.html). It gives me the following:

`pymars -help
usage: pymars [-h] -m MODEL [-e ERROR] [--method {DRG,DRGEP,PFA}]
              [--targets TARGETS [TARGETS ...]] [--conditions CONDITIONS]
              [--retained_species [RETAINED_SPECIES [RETAINED_SPECIES ...]]]
              [--sensitivity_analysis] [--sensitivity_type {initial,greedy}]
              [--upper_threshold UPPER_THRESHOLD] [--path PATH]
              [--num_threads [NUM_THREADS]] [--convert] [--thermo THERMO]
              [--transport TRANSPORT]

pyMARS: Reduce chemical kinetic models.

optional arguments:
  -h, --help            show this help message and exit
  -m MODEL, --model MODEL
                        input model filename (e.g., "mech.cti").
  -e ERROR, --error ERROR
                        Maximum error percentage for the reduced model.
  --method {DRG,DRGEP,PFA}
                        skeletal reduction method to use.
  --targets TARGETS [TARGETS ...]
                        List of target species (e.g., "CH4 O2").
  --conditions CONDITIONS
                        File with list of autoignition initial conditions.
  --retained_species [RETAINED_SPECIES [RETAINED_SPECIES ...]]
                        List of non-target species to always retain (e.g., "N2
                        Ar")
  --sensitivity_analysis
                        Run sensitivity analysis after completing another
                        method.
  --sensitivity_type {initial,greedy}
                        Sensitivity analysis algorithm to be used
  --upper_threshold UPPER_THRESHOLD
                        Upper threshold value used to determine species for
                        sensitivity analysis.
  --path PATH           Path to directory for writing files.
  --num_threads [NUM_THREADS]
                        Number of CPU cores to use for running simulations in
                        parallel. If no number, then use available number of
                        cores minus 1.
  --convert             Convert files between Cantera and Chemkin formats
                        (.cti <=> .inp)
  --thermo THERMO       thermodynamic data filename (only necessary for
                        Chemkin files).
  --transport TRANSPORT
                        transport data filename (only necessary for Chemkin
                        files).
`

And I'm getting the following error. Not sure if it relates to my configuration.

`pymars --input example_input_file.yaml
usage: pymars [-h] -m MODEL [-e ERROR] [--method {DRG,DRGEP,PFA}]
              [--targets TARGETS [TARGETS ...]] [--conditions CONDITIONS]
              [--retained_species [RETAINED_SPECIES [RETAINED_SPECIES ...]]]
              [--sensitivity_analysis] [--sensitivity_type {initial,greedy}]
              [--upper_threshold UPPER_THRESHOLD] [--path PATH]
              [--num_threads [NUM_THREADS]] [--convert] [--thermo THERMO]
              [--transport TRANSPORT]
pymars: error: the following arguments are required: -m/--model`

I have tried to specify the same
