### 2019-08-07 #1543 @kyleniemeyer
Hi @fcontino, it looks like the version you installed is not the latest version (v1.1.0a8). That command-line interface was used before we made some changes in response to @jcsutherland's feedback.

If you installed using conda, could you try installing the latest version with `conda install -c niemeyer-research-group pymars=1.1.0a8`?
