### 2019-08-21 992a85a
Catch ct.CanteraError when the cvodes integrator fails in sensitivity analysis
