### 2019-09-03 #1543 @fcontino
Sorry for the huge delay. I'm still having issues. Can you help me?

conda install -c niemeyer-research-group pymars=1.1.0a8
Solving environment: failed
Initial quick solve with frozen env failed.  Unfreezing env and trying again.
Solving environment: failed

UnsatisfiableError: The following specifications were found to be incompatible with each other:

```
Package entrypoints conflicts for:
_ipyw_jlab_nb_ext_conf -> jupyterlab -> jupyterlab_server[version='>=0.2.0,<0.3.0'] -> notebook[version='>=4.2.0'] -> nbconvert -> entrypoints[version='>=0.2.2']
jupyterlab -> jupyterlab_server[version='>=0.2.0,<0.3.0'] -> notebook[version='>=4.2.0'] -> nbconvert -> entrypoints[version='>=0.2.2']
spyder -> nbconvert -> entrypoints[version='>=0.2.2']
entrypoints
nbconvert -> entrypoints[version='>=0.2.2']
jupyter -> notebook -> nbconvert -> entrypoints[version='>=0.2.2']
anaconda==5.2.0=py36_3 -> jupyterlab==0.32.1=py36_0 -> notebook[version='>=4.2.0'] -> nbconvert -> entrypoints[version='>=0.2.2']
nltk -> twython -> requests-oauthlib[version='>=0.4.0'] -> oauthlib[version='>=0.6.2'] -> pyjwt[version='>=1.0.0'] -> pep8-naming -> flake8-polyfill[version='>=1.0.2,>=1.0.2,<2'] -> flake8 -> entrypoints[version='>=0.3.0,<0.4.0']
widgetsnbextension -> notebook[version='>=4.4.1'] -> nbconvert -> entrypoints[version='>=0.2.2']
notebook -> nbconvert -> entrypoints[version='>=0.2.2']
jupyterlab_launcher -> notebook -> nbconvert -> entrypoints[version='>=0.2.2']
Package pyyaml conflicts for:
anaconda-navigator -> anaconda-project[version='>=0.4'] -> anaconda-client -> pyyaml[version='>=3.12']
anaconda-client -> pyyaml[version='>=3.12']
odo -> dask[version='>=0.11.1'] -> bokeh[version='>=0.13.0'] -> pyyaml[version='>=3.10']
conda-verify -> pyyaml
pyyaml
dask -> bokeh[version='>=0.13.0'] -> pyyaml[version='>=3.10']
bokeh -> pyyaml[version='>=3.10']
distributed -> pyyaml
conda-build -> conda-verify -> pyyaml
anaconda==5.2.0=py36_3 -> blaze==0.11.3=py36h02e7a37_0 -> odo[version='>=0.5.0'] -> dask[version='>=0.11.1'] -> bokeh -> pyyaml[version='>=3.10']
anaconda-project -> anaconda-client -> pyyaml[version='>=3.12']
scikit-image -> dask[version='>=0.5'] -> bokeh[version='>=0.13.0'] -> pyyaml[version='>=3.10']
blaze -> odo[version='>=0.5.0'] -> dask[version='>=0.11.1'] -> bokeh[version='>=0.13.0'] -> pyyaml[version='>=3.10']
pymars=1.1.0a8 -> pyyaml[version='>=4.2b1']
Package mkl-service conflicts for:
mkl-service
```
