### 2019-09-03 #1543 @katyhuff
Thank you @jcsutherland and @fcontino for your reviews -- we couldn't do this without you.
Also, of course, thank you @kyleniemeyer for a strong submission and for engaging actively in the review process!

I have looked over the paper, double-checked all the DOI links, and have conducted a high-level review of the code itself. The only thing that renders strangely is the pytables citation (but I recognize that this is a feature of their chosen CITATION.bib file). Since there's not much to be done about that, and everything looks ship-shape to me, I think we are ready to move forward with accepting this submission.

@kyleniemeyer : At this point, please double-check the paper yourself, review any lingering details in your code/readme/etc., and then make an archive of the reviewed software in Zenodo/figshare/other service. Please be sure that the DOI metadata (title, authors, etc.) matches this JOSS submission. Once that's complete, please update this thread with the DOI of the archive, and I'll move forward with accepting the submission. Until then, now is your moment for final touchups!
