### 2019-09-06 9c5716a
updated JOSS paper
