### 2019-09-06 b568baa
:gem: v1.1.0 release
