### 2019-09-06 c630c02
trying to fix pytables reference in paper
