### 2019-09-06 50ce1d4
(still) trying to fix pytables reference in paper [skip ci]
