+++ paper/paper.bib
@@ -214,7 +214,7 @@
 @Misc{pytables,
   author =    {{PyTables Developers Team}},
   title =     {{PyTables}: Hierarchical Datasets in {Python}},
-  year =      {2002--present},
+  year =      {2002-present},
   url = {https://www.pytables.org/}
 }
 
