### 2019-09-06 06dafae
(still) trying to fix pytables reference in paper [skip ci]
