### 2019-09-06 #1543 @kyleniemeyer
@katyhuff ok, sorry for the spam—I resolved that pytables reference by just changing it to "2002-2019", which is still accurate.

I released v1.1.0, and the DOI is 10.5281/zenodo.3401549.

Thanks everyone!
