### 2019-09-07 #1543 @fcontino
Congrats Kyle! Nice work!

On Fri, Sep 6, 2019 at 6:59 PM Kyle Niemeyer <notifications@github.com>
wrote:

> @katyhuff <https://github.com/katyhuff> ok, sorry for the spam—I resolved
> that pytables reference by just changing it to "2002-2019", which is still
> accurate.
>
> I released v1.1.0, and the DOI is 10.5281/zenodo.3401549.
>
> Thanks everyone!
>
