### 2019-09-07 #1543 @katyhuff
Congratulations @kyleniemeyer, and thanks again @jcsutherland and @fcontino for your reviews!

@openjournals/jose-eics I think this is ready for acceptance!
